# server/docs/swagger.yaml
components:
  schemas:
    api.ErrorResponse:
      properties:
        code:
          example: 400
          type: integer
        error:
          example: validation failed
          type: string
        message:
          example: Bad request
          type: string
      type: object
...
    dbtypes.AudioSpecificMetadata:
      properties:
        album:
          example: Album Title
          type: string
        artist:
          example: John Doe
          type: string
        bitrate:
          example: 128000
          type: integer
        channels:
          example: 2
          type: integer
        codec:
          example: AAC
          type: string
        description:
          example: Song Description
          type: string
        genre:
          example: Pop
          type: string
        sample_rate:
          example: 44100
          type: integer
        title:
          example: Song Title
          type: string
        year:
          example: 2023
          type: integer
      type: object
    dbtypes.PhotoSpecificMetadata:
      properties:
        camera_model:
          type: string
        capture_offset_minutes:
          type: integer
        content_identifier:
          type: string
        description:
          type: string
        dimensions:
          type: string
        exposure:
          type: number
        exposure_time:
          type: string
        f_number:
          type: number
        focal_length:
          type: number
        gps_latitude:
          type: number
        gps_longitude:
          type: number
        is_raw:
          type: boolean
        iso_speed:
          type: integer
        lens_model:
          type: string
        resolution:
          type: string
        taken_time:
          type: string
      type: object
    dbtypes.SpeciesPredictionMeta:
      properties:
        label:
          description: 物种标签
          type: string
        score:
          description: 置信度分数
          type: number
      type: object
    dbtypes.VideoSpecificMetadata:
      properties:
        bitrate:
          example: 1000000
          type: integer
        camera_model:
          example: Canon EOS 5D Mark IV
          type: string
        capture_offset_minutes:
          type: integer
        codec:
          example: H.264
          type: string
        content_identifier:
          type: string
        description:
          example: A beautiful sunset over the ocean
          type: string
        frame_rate:
          example: 30
          type: number
        gps_latitude:
          example: 37.7749
          type: number
        gps_longitude:
          example: -122.4194
          type: number
        recorded_time:
          example: "2023-01-01T00:00:00Z"
          type: string
      type: object
...
    dto.AgentDateRangeDTO:
      properties:
        from:
          type: string
        to:
          type: string
      type: object
    dto.AgentFacetBucket:
      properties:
        bucket:
          example: 2025-04
          type: string
        count:
          example: 12
          type: integer
      type: object
    dto.AgentNameCountDTO:
      properties:
        count:
          example: 42
          type: integer
        name:
          example: Kyoto
          type: string
      type: object
...
    dto.AgentRefAssetsDTO:
      properties:
        assets:
          items:
            $ref: '#/components/schemas/dto.AssetDTO'
          type: array
          uniqueItems: false
        pagination:
          $ref: '#/components/schemas/dto.PaginationDTO'
        total:
          example: 97
          type: integer
      type: object
    dto.AgentRefDTO:
      properties:
        count:
          example: 97
          type: integer
        created_at:
          type: string
        facets:
          $ref: '#/components/schemas/dto.AgentRefFacetsDTO'
        op:
          example: combine
          type: string
        ref_id:
          example: r3_kyoto
          type: string
        truncated:
          type: boolean
      type: object
    dto.AgentRefFacetsDTO:
      properties:
        cameras:
          items:
            $ref: '#/components/schemas/dto.AgentNameCountDTO'
          type: array
          uniqueItems: false
        count:
          type: integer
        date_range:
          $ref: '#/components/schemas/dto.AgentDateRangeDTO'
        histogram:
          items:
            $ref: '#/components/schemas/dto.AgentFacetBucket'
          type: array
          uniqueItems: false
        liked_count:
          type: integer
        rating_dist:
          items:
            type: integer
          type: array
          uniqueItems: false
        top_people:
          items:
            $ref: '#/components/schemas/dto.AgentNameCountDTO'
          type: array
          uniqueItems: false
        top_places:
          items:
            $ref: '#/components/schemas/dto.AgentNameCountDTO'
          type: array
          uniqueItems: false
        types:
          additionalProperties:
            type: integer
          type: object
      type: object
    dto.AssetDTO:
      properties:
        asset_id:
          type: string
        capture_offset_minutes:
          type: integer
        deleted_at:
          type: string
        duration:
          type: number
        file_size:
          type: integer
        hash:
          type: string
        height:
          type: integer
        is_deleted:
          type: boolean
        liked:
          type: boolean
        mime_type:
          type: string
        original_filename:
          type: string
        owner_id:
          type: integer
        rating:
          type: integer
        repository_id:
          type: string
        species_predictions:
          items:
            $ref: '#/components/schemas/dbtypes.SpeciesPredictionMeta'
          type: array
          uniqueItems: false
        specific_metadata:
          oneOf:
          - $ref: '#/components/schemas/dbtypes.PhotoSpecificMetadata'
          - $ref: '#/components/schemas/dbtypes.VideoSpecificMetadata'
          - $ref: '#/components/schemas/dbtypes.AudioSpecificMetadata'
          type: object
        stack:
          $ref: '#/components/schemas/dto.StackPreviewDTO'
        status:
          items:
            type: integer
          type: array
          uniqueItems: false
        storage_path:
          type: string
        taken_time:
          type: string
        type:
          type: string
        upload_time:
          type: string
        width:
          type: integer
      type: object
...
    dto.PaginationDTO:
      description: limit, offset
      properties:
        limit:
          example: 20
          maximum: 100
          minimum: 1
          type: integer
        offset:
          example: 0
          minimum: 0
          type: integer
      type: object
...
    dto.StackPreviewDTO:
      description: Stack fields (populated when stack mode is enabled)
      properties:
        stack_cover:
          description: Whether this asset is the cover of its stack
          example: true
          type: boolean
        stack_id:
          example: 550e8400-e29b-41d4-a716-446655440000
          type: string
        stack_kind:
          enum:
          - raw_jpeg
          - live_photo
          - manual
          example: live_photo
          type: string
        stack_size:
          description: Number of members in the stack
          example: 3
          type: integer
      type: object
...
    handler.AgentChatRequest:
      properties:
        context:
          items:
            $ref: '#/components/schemas/inject.ContextItem'
          type: array
          uniqueItems: false
        mentions:
          items:
            $ref: '#/components/schemas/inject.MentionItem'
          type: array
          uniqueItems: false
        query:
          type: string
        thread_id:
          type: string
      required:
      - query
      type: object
    handler.AgentResumeRequest:
      properties:
        targets:
          additionalProperties: {}
          type: object
        thread_id:
          type: string
      required:
      - targets
      - thread_id
      type: object
...
    inject.ContextItem:
      properties:
        asset_ids:
          items:
            type: string
          type: array
          uniqueItems: false
        label:
          type: string
        type:
          type: string
      type: object
    inject.MentionItem:
      properties:
        id:
          type: string
        label:
          type: string
        type:
          type: string
      type: object
  securitySchemes:
    BearerAuth:
      description: Type "Bearer" followed by a space and JWT token.
      in: header
      name: Authorization
      type: apiKey
...
info:
  contact:
    name: API Support
    url: http://www.github.com/EdwinZhanCN/Lumilio-Photos
  description: Media management system API with asset features
  license:
    name: GPLv3.0
    url: https://opensource.org/licenses/GPL-3.0
  title: Lumilio-Photos API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /api/v1/agent/chat:
    post:
      description: Send a query to agent and receive streaming responses via SSE.
        Manages conversation threads.
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/handler.AgentChatRequest'
                description: Chat request
                summary: request
        description: Chat request
        required: true
      responses:
        "200":
          content:
            text/event-stream:
              schema:
                type: string
          description: SSE stream
        "400":
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Invalid request
        "401":
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unauthorized
        "500":
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal server error
      summary: Chat with Agent
      tags:
      - agent
  /api/v1/agent/chat/resume:
    post:
      description: Resume a conversation from an interrupt point (e.g., user confirmation
        for a tool call)
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/handler.AgentResumeRequest'
                description: Resume request
                summary: request
        description: Resume request
        required: true
      responses:
        "200":
          content:
            text/event-stream:
              schema:
                type: string
          description: SSE stream
        "400":
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Invalid request
        "401":
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unauthorized
        "500":
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal server error
      summary: Resume Agent Chat
      tags:
      - agent
...
  /api/v1/agent/refs/{id}:
    get:
      description: Get metadata and facet summary for an agent ref. Refs are scoped
        to the requesting user and thread.
      parameters:
      - description: Ref ID
        in: path
        name: id
        required: true
        schema:
          type: string
      - description: Thread (conversation) the ref belongs to
        in: query
        name: thread_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.AgentRefDTO'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Ref not found
      summary: Get Agent Ref Metadata
      tags:
      - agent
  /api/v1/agent/refs/{id}/assets:
    get:
      description: Get a page of assets for an agent ref, in snapshot order.
      parameters:
      - description: Ref ID
        in: path
        name: id
        required: true
        schema:
          type: string
      - description: Thread (conversation) the ref belongs to
        in: query
        name: thread_id
        required: true
        schema:
          type: string
      - description: Page size (default 50, max 200)
        in: query
        name: limit
        schema:
          type: integer
      - description: Page offset (default 0)
        in: query
        name: offset
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.AgentRefAssetsDTO'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Ref not found
      summary: Get Agent Ref Assets
      tags:
      - agent
...
servers:
- url: localhost:6680/api/v1